chart updates

+++ database/migrations/2023_03_11_185234_create_chart_data_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->integer('customer_id');
             $table->integer('chart_id');
-            $table->integer('weight');
-            $table->integer('fat_percentage');
+            $table->float('weight');
+            $table->float('fat_percentage');
             $table->timestamps();
         });
     }
